Release v1.4.0
Acquiring server-side helpers, money UAH constructors, and a signed
webhook generator for tests. All additive — no breaking changes.

Added:
- acquiring.NewWebhookHandler: ready http.Handler with key auto-refresh,
  freshness check, and pluggable dedup by (invoiceId, modifiedDate).
- acquiring reconciliation: Client.PollInvoice, ReconcileStatement,
  InvoiceStatus.IsTerminal.
- acquiring basket builder: NewBasket/NewBasketItem with validation.
- acquiring typed errors: APIError + Is{NotFound,TooManyRequests,...}.
- acquiring subscription lifecycle: ClassifyCharge/ClassifySubscription.
- money.UAH/FromMajor/ParseMajor.
- monobanktest.NewAcquiringWebhookSigner + Server.WithAcquiringPubKey.

Changed:
- Migrated golangci-lint config to v2; fixed builtin max shadowing in
  retry/webhook; US-English spelling consistency.

+++ corporate/range_test.go
@@ -43,7 +43,7 @@ func TestTransactionsRange_paginates31DayWindows(t *testing.T) {
 
 	// Chronological order: each window's from < next window's from.
 	for i := 1; i < len(got); i++ {
-		assert.Less(t, int64(got[i-1].Time.Unix()), int64(got[i].Time.Unix()))
+		assert.Less(t, got[i-1].Time.Unix(), got[i].Time.Unix())
 	}
 }
